Somewhat working
I have gotten the code to work somewhat, but currently it will load the page and all is showing, but when you click on the tabs they are not showing the specific ones, until you select all of the tabs then click on them again. After going through all the tabs then the specific articles for the specific tab will come up, but whats weird is when you press all nothing will show up.