Complete carousel functionality
I went through the first part of the stretch task to complete the carousel. Because I thought it was important and not too challenging I added the infinite loop as part of the base functionality. The code is a little wet, but I was unsure if things would work so I kept it simple. If there is time, I will go back and dry up the code a bit.