Refactor carousel
I began work on adding an animation to the carousel, but I was having issue with less-watch-compiler. It refused to automatically update any saved changes, so tinkering with CSS was taking forever. There was no time during this challenge to debug less-watch-compiler, so I used my time to refactor the code so it was not as wet.